rename file observatory_discovery -> mdns_discovery
