+++ packages/flutter_tools/lib/src/ios/devices.dart
@@ -19,9 +19,9 @@ import '../build_info.dart';
 import '../convert.dart';
 import '../device.dart';
 import '../globals.dart';
-import '../observatory_discovery.dart';
-import '../protocol_discovery.dart';
+import '../mdns_discovery.dart';
 import '../project.dart';
+import '../protocol_discovery.dart';
 import '../reporting/reporting.dart';
 import 'code_signing.dart';
 import 'ios_workflow.dart';
